requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.9.2
annotated-types==0.7.0; python_version >= '3.8'
cachetools==5.4.0; python_version >= '3.7'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_version >= '3.7'
click==8.1.3
flask-cors==3.0.10
flask-login==0.6.2
flask-migrate==4.0.2
flask-sqlalchemy==3.0.2
flask-wtf==1.1.1
flask==2.2.2
google-ai-generativelanguage==0.6.6; python_version >= '3.7'
google-api-core==2.19.1; python_version >= '3.7'
google-api-python-client==2.139.0; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-auth==2.32.0; python_version >= '3.7'
google-generativeai==0.7.2
googleapis-common-protos==1.63.2; python_version >= '3.7'
greenlet==3.0.1
grpcio-status==1.62.2
grpcio==1.65.4
gunicorn==20.1.0
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.7; python_version >= '3.5'
importlib-metadata==6.9.0
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.2.4
markupsafe==2.1.2
proto-plus==1.24.0; python_version >= '3.7'
protobuf==4.25.4; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pyasn1==0.6.0; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pydantic==2.8.2; python_version >= '3.8'
pyparsing==3.1.2; python_version >= '3.1'
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
requests==2.32.3; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==72.1.0; python_version >= '3.8'
six==1.16.0
sqlalchemy==1.4.46
tqdm==4.66.5; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.2; python_version >= '3.8'
werkzeug==2.2.2
wtforms==3.0.1
zipp==3.17.0
